detail product

+++ database/migrations/2024_07_11_034240_create_products_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->text('deskripsi_produk')->nullable();
             $table->string('merek')->nullable();
             $table->float('berat')->nullable();
-            $table->float('harga_jual')->nullable();
+            $table->decimal('harga_jual', 15, 2)->nullable();
             $table->integer('total_stock')->nullable();
             $table->enum('status', ['tersedia', 'habis'])->default('tersedia')->nullable();
             $table->string('gambar')->nullable();
